coordinator queries contract addresses using ownership
